Moved cache decorators tests into decorators/test_cache.py.
